Simplified a lot of stuff
Moved the try catch blocks of all the load methods to within the
methods. Changed how the updateAds method works to make it easier to
comprehend and less messy, as well as more efficient. Removed the now
unnecessary compareTo method in Ad, and removed a lot of accessor
methods that were made unnecessary when I merged the old Craigslist.java
with CraigslistNotifier.java.