Changed back to updating from html instead of rss
The RSS feed doesn't refresh very often (every hour or two it seems),
and so retrieving ads from it introduced an unnecessary delay.
Unfortunately this means the emails can't contain the body of the ad
because that would take too many requests to CL's servers, but its worth
that tradeoff.