Few changes
Previously, if an ad's html didn't contain a location, the new ad would
just use the previous ad's location, resulting in the user being
notified of the same ad multiple times if it didn't have a location.
Also made some of the substrings more specific to ensure they would
parse the html correctly.